Cleaned up the Redux implementation. Included all proper TS types.

+++ frontend/src/Components/MovieCards/MovieStatusCard.tsx
@@ -1,5 +1,5 @@
 import {useDispatch, useSelector} from "react-redux";
-import {set, UserState} from "../../Store/userSlice";
+import {RootState, set, UserState} from "../../Store/userSlice";
 import React, {useEffect, useState} from "react";
 
 const status_types = [
@@ -11,27 +11,33 @@ const status_types = [
 ]
 // TODO Change user variable name to be pragmatic
 export function MovieStatusCard(props: { movieID: number }) {
-    const user = useSelector((state: { user: UserState }) => state.user);
+    const user = useSelector<RootState, UserState>((state) => state.user);
     const dispatch = useDispatch()
     const [loading, setLoading] = useState<boolean>(false)
     const [curMovieStatus, setCurMovieStatus] = useState<string | null>(null)
 
     useEffect(() => {
-        let matchingShow = user.user.movieList.filter(obj => {
-            return obj.id === props.movieID
-        })
-        if (matchingShow.length > 0) {
-            setCurMovieStatus(matchingShow[0].status)
+        if(user.profile !== null){
+            let matchingShow = user.profile.movieList.filter(obj => {
+                return obj.id === props.movieID
+            })
+            if (matchingShow.length > 0) {
+                setCurMovieStatus(matchingShow[0].status)
+            }
         }
+
     }, [user]);
 
 
     function addMovieToWatchlist(id: number, status: string) {
+        if (user.profile == null){
+            return
+        }
         setLoading(true)
         fetch(process.env.REACT_APP_HOST + "/api/v1/movie/" + id + "/" + status, {
             method: "PUT",
             headers: {
-                'AuthToken': user.user.idToken
+                'AuthToken': user.profile.idToken
             }
         })
             .then((res) => {
@@ -41,21 +47,21 @@ export function MovieStatusCard(props: { movieID: number }) {
             })
             .then(
                 (result) => {
-                    let tempUser = JSON.parse(JSON.stringify(user));
+                    let tempUserProfile = JSON.parse(JSON.stringify(user.profile));
                     let index = -1;
-                    for (let i = 0; i < tempUser.user.movieList.length; i++) {
-                        if (tempUser.user.movieList[i].id === result.id) {
+                    for (let i = 0; i < tempUserProfile.movieList.length; i++) {
+                        if (tempUserProfile.movieList[i].id === result.id) {
                             index = i;
                             break;
                         }
                     }
                     if (index > -1) {
-                        tempUser.user.movieList[index] = result
+                        tempUserProfile.movieList[index] = result
                     } else {
-                        tempUser.user.movieList.unshift(result)
+                        tempUserProfile.movieList.unshift(result)
                     }
 
-                    dispatch(set(tempUser.user))
+                    dispatch(set(tempUserProfile))
                     setLoading(false)
                 }, (error) => {
                     setLoading(false)
@@ -64,10 +70,13 @@ export function MovieStatusCard(props: { movieID: number }) {
     }
 
     function deleteFromWatchlist(id: number) {
+        if (user.profile == null){
+            return
+        }
         fetch(process.env.REACT_APP_HOST + "/api/v1/movie/" + id, {
             method: "DELETE",
             headers: {
-                'AuthToken': user.user.idToken
+                'AuthToken': user.profile.idToken
             }
         })
             .then((res) => {
@@ -93,7 +102,7 @@ export function MovieStatusCard(props: { movieID: number }) {
             )
     }
 
-    if (!user.userExists) {
+    if (user === null) {
         return <></>
     }
     return (
